reimplement the fast but O(nm) space alignment algorithm
All tests are unchanged and passed.
